fix seo meta tags
