+++ src/app/ahif-2025/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
@@ -8,8 +7,10 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from "@/components/ui/accordion";
-import { useState, useEffect } from "react";
 import Header from "./components/header";
+import { Metadata } from "next";
+import AHIFNavigation from "./components/navigation";
+import AHIFHeroImage from "@/assets/images/ahif/ahif-hero-image.png";
 
 const modules = [
   {
@@ -43,174 +44,13 @@ const modules = [
   },
 ];
 
-export default function AHIF2025() {
-  const [activeSection, setActiveSection] = useState("overview");
-  const [isScrolled, setIsScrolled] = useState(false);
-
-  useEffect(() => {
-    const observerOptions = {
-      root: null,
-      rootMargin: "0px",
-      threshold: 0.6,
-    };
-
-    const observerCallback = (entries: any[]) => {
-      entries.forEach((entry) => {
-        if (entry.isIntersecting) {
-          setActiveSection(entry.target.id);
-        }
-      });
-    };
-
-    const observer = new IntersectionObserver(
-      observerCallback,
-      observerOptions
-    );
-
-    // Observe all sections
-    const sections = document.querySelectorAll("section[id]");
-    sections.forEach((section) => observer.observe(section));
-
-    // Handle scroll for sticky nav
-    const handleScroll = () => {
-      const scrollPosition = window.scrollY;
-      if (scrollPosition > 100) {
-        setIsScrolled(true);
-      } else {
-        setIsScrolled(false);
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      sections.forEach((section) => observer.unobserve(section));
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
-  const scrollToSection = (sectionId: string) => {
-    const section = document.getElementById(sectionId);
-    const navHeight = document.getElementById("sticky-nav")?.offsetHeight || 0;
-
-    if (section) {
-      const sectionTop =
-        section.getBoundingClientRect().top + window.pageYOffset;
-      window.scrollTo({
-        top: sectionTop - navHeight - 20, // Subtract nav height and add some padding
-        behavior: "smooth",
-      });
-    }
-  };
-
+export default function AHIF2025Page() {
   return (
     <>
       <Header />
 
       {/* Sticky Navigation */}
-      <div
-        id="sticky-nav"
-        className={cn(
-          "sticky top-0 w-full bg-white z-50 transition-all duration-300",
-          isScrolled ? "shadow-md py-3" : "py-4"
-        )}
-      >
-        <div className="max-w-screen-xl mx-auto px-4">
-          <div className="overflow-x-auto pb-2 hide-scrollbar">
-            <div className="flex gap-2 md:gap-4 min-w-max">
-              <Link
-                className={cn(
-                  "text-primary font-medium py-2 px-3 md:py-3 md:px-5 rounded-md transition-colors whitespace-nowrap text-sm md:text-base",
-                  {
-                    "bg-primary-green/10 hover:bg-primary-green/20 text-primary-green":
-                      activeSection !== "overview",
-                    "bg-primary-green text-white": activeSection === "overview",
-                  }
-                )}
-                href="#overview"
-                onClick={(e) => {
-                  e.preventDefault();
-                  scrollToSection("overview");
-                }}
-              >
-                Overview
-              </Link>
-              <Link
-                className={cn(
-                  "text-primary font-medium py-2 px-3 md:py-3 md:px-5 rounded-md transition-colors whitespace-nowrap text-sm md:text-base",
-                  {
-                    "bg-primary-green/10 hover:bg-primary-green/20 text-primary-green":
-                      activeSection !== "structure",
-                    "bg-primary-green text-white":
-                      activeSection === "structure",
-                  }
-                )}
-                href="#structure"
-                onClick={(e) => {
-                  e.preventDefault();
-                  scrollToSection("structure");
-                }}
-              >
-                Structure
-              </Link>
-              <Link
-                className={cn(
-                  "text-primary font-medium py-2 px-3 md:py-3 md:px-5 rounded-md transition-colors whitespace-nowrap text-sm md:text-base",
-                  {
-                    "bg-primary-green/10 hover:bg-primary-green/20 text-primary-green":
-                      activeSection !== "core-modules",
-                    "bg-primary-green text-white":
-                      activeSection === "core-modules",
-                  }
-                )}
-                href="#core-modules"
-                onClick={(e) => {
-                  e.preventDefault();
-                  scrollToSection("core-modules");
-                }}
-              >
-                Core Module
-              </Link>
-              <Link
-                className={cn(
-                  "text-primary font-medium py-2 px-3 md:py-3 md:px-5 rounded-md transition-colors whitespace-nowrap text-sm md:text-base",
-                  {
-                    "bg-primary-green/10 hover:bg-primary-green/20 text-primary-green":
-                      activeSection !== "why-apply",
-                    "bg-primary-green text-white":
-                      activeSection === "why-apply",
-                  }
-                )}
-                href="#why-apply"
-                onClick={(e) => {
-                  e.preventDefault();
-                  scrollToSection("why-apply");
-                }}
-              >
-                Why Apply
-              </Link>
-              <Link
-                className={cn(
-                  "text-primary font-medium py-2 px-3 md:py-3 md:px-5 rounded-md transition-colors whitespace-nowrap text-sm md:text-base",
-                  {
-                    "bg-primary-green/10 hover:bg-primary-green/20 text-primary-green":
-                      activeSection !== "eligibility",
-                    "bg-primary-green text-white":
-                      activeSection === "eligibility",
-                  }
-                )}
-                href="#eligibility"
-                onClick={(e) => {
-                  e.preventDefault();
-                  scrollToSection("eligibility");
-                }}
-              >
-                Eligibility
-              </Link>
-            </div>
-          </div>
-        </div>
-      </div>
+      <AHIFNavigation />
 
       {/* Introduction */}
       <section id="overview" className="bg-white py-14 lg:py-20">
@@ -558,3 +398,35 @@ export default function AHIF2025() {
     </>
   );
 }
+
+export const metadata: Metadata = {
+  title: "Africa Healthcare Innovation Fellowship 2025",
+  description:
+    "Empowering innovators to transform healthcare in Africa through digital solutions.",
+  alternates: {
+    canonical: "https://innovatehealth.africa/ahif-2025",
+  },
+
+  openGraph: {
+    title: "Africa Healthcare Innovation Fellowship 2025",
+    description:
+      "Empowering innovators to transform healthcare in Africa through digital solutions.",
+    url: "https://innovatehealth.africa/ahif-2025",
+    type: "website",
+    images: [
+      {
+        url: AHIFHeroImage.src,
+        width: 1200,
+        height: 630,
+        alt: "AHIF 2025 Banner",
+      },
+    ],
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "Africa Healthcare Innovation Fellowship 2025",
+    description:
+      "Empowering innovators to transform healthcare in Africa through digital solutions.",
+    images: [AHIFHeroImage.src],
+  },
+};
